PROD-722 Post visibility와 Activity 소유 경계를 정리한다
