+++ apps/worker/package.json
@@ -16,7 +16,6 @@
     "@kosmo/fedify": "workspace:*",
     "@temporalio/worker": "^1.22.0",
     "@temporalio/workflow": "^1.22.0",
-    "drizzle-orm": "1.0.0-beta.22",
     "tsx": "^4.22.4"
   }
 }
