+++ openspec/changes/transition-post-create-to-temporal-command-workflow/design.md
@@ -42,7 +42,7 @@ Worker foundation은 health·signal 경계를 제공하지만 production entrypo
 
 Workflow input은 committed Post ID와 origin처럼 다시 조회 가능한 최소 identity만 포함한다. 같은 Post에 대한 반복 start는 Temporal의 Workflow ID conflict policy로 실행 중인 accepted execution에 수렴하고 reuse policy로 종료된 execution의 재시작을 거부한다. duplicate/no-op Post 결과에서는 start를 시도하지 않는다. Start가 accepted되기 전의 누락은 복구하지 않는다.
 
-Accepted Workflow는 Reply Notification과 Local-origin Fedify queue handoff를 별도 Activity로 실행한다. 두 효과가 모두 적용될 수 있는 경우 서로 독립적으로 시작하고 각 Activity의 retry/최종 실패가 다른 효과를 선행 차단하지 않게 결과를 함께 수집한다. Notification Activity는 process 기본 `db`로 committed Post를 다시 조회해 기존 recipient·self suppression·visibility·uniqueness 정책과 멱등 insert를 직접 소유한다. production caller가 하나뿐인 별도 core service wrapper는 두지 않는다. Delivery Activity는 기존 canonical Local Note/Activity identity, audience/target과 queue producer를 재사용하며 `origin=ACTIVITYPUB`이면 생성하지 않는다. Local Delete가 먼저 commit되면 기존 Delete handoff가 삭제된 Post의 보존된 content projection으로 Create를 먼저 enqueue하고 같은 Note ordering key로 Delete를 이어서 enqueue한다. Create projection 조회 뒤 queue handoff 중 Delete가 commit되는 경쟁에서는 DB row lock을 외부 queue I/O까지 유지하지 않고, Create handoff 뒤 삭제 상태를 다시 확인해 필요한 Create/Delete 보정 쌍을 같은 ordering key에 덧붙인다. Fedify PostgreSQL MessageQueue는 dedup key를 제공하지 않으므로 이 보정과 Temporal retry가 queue-level exactly-once를 보장하지는 않는다. acknowledgement가 모호하면 같은 Activity ID의 중복 queue message나 HTTP request가 가능하고, 수신 측 ActivityPub idempotency로 의미상 수렴한다. 이를 제거하려면 현재 범위에서 금지한 queue schema 변경이나 별도 delivery history가 필요하다.
+Accepted Workflow는 Reply Notification과 Local-origin Fedify queue handoff를 별도 Activity로 실행한다. 두 효과가 모두 적용될 수 있는 경우 서로 독립적으로 시작하고 각 Activity의 retry/최종 실패가 다른 효과를 선행 차단하지 않게 결과를 함께 수집한다. Reply Notification materialization은 Temporal SDK나 Worker runtime state에 의존하지 않는 transport-neutral state change이므로 전용 Core service 파일의 함수 하나로 둔다. 이 함수가 process 기본 `db`로 committed Post를 다시 조회하고 공통 Post visibility policy를 재사용해 기존 recipient·self suppression·visibility·uniqueness 정책과 멱등 insert를 직접 소유한다. Worker는 별도 wrapper나 Activity 전용 Core 계층 없이 같은 함수를 Activity 이름으로 alias해 직접 등록한다. Delivery Activity는 기존 canonical Local Note/Activity identity, audience/target과 queue producer를 재사용하며 `origin=ACTIVITYPUB`이면 생성하지 않는다. Local Delete가 먼저 commit되면 기존 Delete handoff가 삭제된 Post의 보존된 content projection으로 Create를 먼저 enqueue하고 같은 Note ordering key로 Delete를 이어서 enqueue한다. Create projection 조회 뒤 queue handoff 중 Delete가 commit되는 경쟁에서는 DB row lock을 외부 queue I/O까지 유지하지 않고, Create handoff 뒤 삭제 상태를 다시 확인해 필요한 Create/Delete 보정 쌍을 같은 ordering key에 덧붙인다. Fedify PostgreSQL MessageQueue는 dedup key를 제공하지 않으므로 이 보정과 Temporal retry가 queue-level exactly-once를 보장하지는 않는다. acknowledgement가 모호하면 같은 Activity ID의 중복 queue message나 HTTP request가 가능하고, 수신 측 ActivityPub idempotency로 의미상 수렴한다. 이를 제거하려면 현재 범위에서 금지한 queue schema 변경이나 별도 delivery history가 필요하다.
 
 Production Worker entrypoint는 compile-time registration을 별도 module 없이 inline으로 소유하고 process-global host를 한 번 시작한다. 별도 `runWorker`/`startWorker` callable lifecycle은 두지 않는다. health server, Temporal connection, Worker와 signal handler도 같은 entrypoint가 단독 소유한다. Helm은 유효한 release image가 주어지는 모든 환경에서 application workloads를 항상 render하고, chart-wide 또는 Worker별 activation key를 사용하지 않으며, Workflow producer인 GraphQL API와 ActivityPub queue consumer에도 같은 환경별 Temporal endpoint·namespace를 전달한다. API에는 Worker database credential을 주입하지 않는다. Worker DB principal·Secret wiring은 이 change에서 바꾸지 않고 표준 process `db`를 소비한다.
