+++ packages/core/package.json
@@ -14,6 +14,7 @@
     "./post-content": "./post-content/index.ts",
     "./post-content/server": "./post-content/server.ts",
     "./polyfill": "./polyfill.ts",
+    "./post-visibility": "./post-visibility.ts",
     "./profile": "./profile.ts",
     "./search": "./search.ts",
     "./services": "./services/index.ts",
